requirements.txt
django
reportlab==4.4.0
google-generativeai==2025.1.31